Add admin check in download_controller.rb and user_controller.rb so that only privileged users can access POST, PUT, DELETE endpoints

+++ app/controllers/api/v1/download_controller.rb
@@ -3,7 +3,6 @@
 module Api
   module V1
     class DownloadController < ApplicationController
-      skip_before_action :verify_authenticity_token
       def index
         app_name = params[:application_name]
         downloads = Download.order('created_at DESC')
@@ -14,13 +13,11 @@ module Api
             filtered_downloads.push(d) if d.application_name == app_name
           end
           render json: {
-            status: 'SUCCESS',
             message: "Retrieved downloads for '#{app_name}'",
             data: filtered_downloads
           }, status: 200
         else
           render json: {
-            status: 'SUCCESS',
             message: 'Loaded downloads',
             data: downloads
           }, status: 200
@@ -30,53 +27,68 @@ module Api
       def show
         download = Download.find(params[:id])
         render json: {
-          status: 'SUCCESS',
           message: 'Loaded download',
           data: download
         }, status: 200
       end
 
       def create
-        download = Download.new(download_params)
-        if download.save
-          render json: {
-            status: 'SUCCESS',
-            message: 'Created download',
-            data: download
-          }, status: 200
+        if current_user
+          download = Download.new(download_params)
+          if download.save
+            render json: {
+              message: 'Created download',
+              data: download
+            }, status: 200
+          else
+            render json: {
+              message: 'Could not create download',
+              data: download
+            }, status: 400
+          end
         else
           render json: {
-            status: 'ERROR',
-            message: 'Could not create download',
-            data: download
-          }, status: 400
+            message: 'Unauthorized',
+            data: nil
+          }, status: 401
         end
       end
 
       def destroy
-        download = Download.find(params[:id])
-        download.destroy
-        render json: {
-          status: 'SUCCESS',
-          message: 'Deleted download',
-          data: download
-        }, status: 200
-      end
-
-      def update
-        download = Download.find(params[:id])
-        if download.update(download_params)
+        if current_user
+          download = Download.find(params[:id])
+          download.destroy
           render json: {
-            status: 'SUCCESS',
-            message: 'Updated download',
+            message: 'Deleted download',
             data: download
           }, status: 200
         else
           render json: {
-            status: 'ERROR',
-            message: 'Could not update download',
-            data: download
-          }, status: 400
+            message: 'Unauthorized',
+            data: nil
+          }, status: 401
+        end
+      end
+
+      def update
+        if current_user
+          download = Download.find(params[:id])
+          if download.update(download_params)
+            render json: {
+              message: 'Updated download',
+              data: download
+            }, status: 200
+          else
+            render json: {
+              message: 'Could not update download',
+              data: download
+            }, status: 400
+          end
+        else
+          render json: {
+            message: 'Unauthorized',
+            data: nil
+          }, status: 401
         end
       end
